requirements.txt
python-telegram-bot==20.8
openai-whisper
torch
ffmpeg-python
